requirements.txt
torch>=2.0.0
transformers>=4.35.0
pyyaml>=6.0.1
numpy>=1.26.0
pytest>=7.4.3
tokenizers>=0.14.1